Extract CoreData helpers into seperate file
